002_games_json_parser

+++ src/main/java/com/slyscrat/impress/service/scheduled/games/SteamApiService.java
@@ -1,6 +1,8 @@
 package com.slyscrat.impress.service.scheduled.games;
 
+import com.slyscrat.impress.model.dto.GameDto;
+
 public interface SteamApiService {
     void checkGamesList();
-    void addGameInfo(long appId);
+    void addGameInfo(int appId);
 }
